# service/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SwarX API",
    "version": "1.0.0",
    "description": "API documentation for SwarX"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Access-Code",
        "description": "Access token for API authentication"
      }
    }
  },
  "paths": {
...
    "/artists": {
      "get": {
        "summary": "Get artist by ID or link",
        "description": "Retrieve artist details using either an artist ID or a JioSaavn artist link",
        "tags": [
          "Artists"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "description": "Artist ID"
          },
          {
            "in": "query",
            "name": "link",
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "JioSaavn artist URL"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "string"
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "songCount",
            "schema": {
              "type": "string"
            },
            "description": "Number of songs to return"
          },
          {
            "in": "query",
            "name": "albumCount",
            "schema": {
              "type": "string"
            },
            "description": "Number of albums to return"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Sort order (ascending or descending)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved artist details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "artist": {
                          "type": "object"
                        },
                        "songs": {
                          "type": "array"
                        },
                        "albums": {
                          "type": "array"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing access token"
          }
        }
      }
    },
    "/artists/{id}": {
      "get": {
        "summary": "Get artist by ID",
        "description": "Retrieve detailed information about a specific artist",
        "tags": [
          "Artists"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Artist ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "string"
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "songCount",
            "schema": {
              "type": "string"
            },
            "description": "Number of songs to return"
          },
          {
            "in": "query",
            "name": "albumCount",
            "schema": {
              "type": "string"
            },
            "description": "Number of albums to return"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Sort order (ascending or descending)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved artist details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "artist": {
                          "type": "object"
                        },
                        "songs": {
                          "type": "array"
                        },
                        "albums": {
                          "type": "array"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing access token"
          },
          "404": {
            "description": "Artist not found"
          }
        }
      }
    },
    "/artists/{id}/songs": {
      "get": {
        "summary": "Get artist songs",
        "description": "Retrieve paginated list of songs by an artist",
        "tags": [
          "Artists"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Artist ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "string"
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Sort order (ascending or descending)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved artist songs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Song"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing access token"
          },
          "404": {
            "description": "Artist not found"
          }
        }
      }
    },
    "/artists/{id}/albums": {
      "get": {
        "summary": "Get artist albums",
        "description": "Retrieve paginated list of albums by an artist",
        "tags": [
          "Artists"
        ],
        "security": [
          {
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Artist ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "string"
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Sort order (ascending or descending)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved artist albums",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "year": {
                            "type": "string"
                          },
                          "songCount": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing access token"
          },
          "404": {
            "description": "Artist not found"
          }
        }
      }
    },
...
  },
...
}